move redis check functions
